+++ node/clients/profile/profileV2.ts
@@ -189,11 +189,11 @@ export class ProfileClientV2 extends JanusClient {
       const addressV2 = address.document
 
       return {
-        addressName: addressV2.name,
-        city: addressV2.localityAreaLevel1,
+        addressName: addressV2.name ?? addressV2.addressName,
+        city: addressV2.locality,
         complement: addressV2.extend,
         country: addressV2.countryCode,
-        geoCoordinates: addressV2.geoCoordinates,
+        geoCoordinates: addressV2.geoCoordinates ?? [],
         id: address.id,
         number: addressV2.streetNumber,
         postalCode: addressV2.postalCode,
@@ -203,7 +203,7 @@ export class ProfileClientV2 extends JanusClient {
         street: addressV2.route,
         userId: addressV2.profileId,
         addressType: addressV2.addressType ?? 'residential',
-        neighborhood: addressV2.neighborhood,
+        neighborhood: addressV2.localityAreaLevel1,
       } as Address
     })
 
@@ -215,9 +215,10 @@ export class ProfileClientV2 extends JanusClient {
           administrativeAreaLevel1: address.state,
           addressType: address.addressType ?? 'residential',
           countryCode: address.country,
-          extend: address.complement ?? '',
-          geoCoordinates: address.geoCoordinates,
-          localityAreaLevel1: address.city,
+          complement: address.complement ?? '',
+          geoCoordinates: address.geoCoordinates ?? [],
+          locality: address.city,
+          localityAreaLevel1: address.neighborhood ?? '',
           name: address.addressName,
           nearly: address.reference ?? '',
           postalCode: address.postalCode,
@@ -225,7 +226,6 @@ export class ProfileClientV2 extends JanusClient {
           route: address.street,
           streetNumber: address.number ?? '',
           receiverName: address.receiverName,
-          neighborhood: address.neighborhood,
         },
       } as AddressV2
     })
@@ -264,7 +264,6 @@ export class ProfileClientV2 extends JanusClient {
           const [address] = addresses.filter(
             (addr) => addr.addressName === addressesV1[0].addressName
           )
-
           if (address) {
             return this.patch(
               `${this.baseUrl}/${profile.id}/addresses/${address.id}`,